Implement Electron
There are a lot of changes here and it's all too messy to split into
separate commits. Honestly a lot of it is over my head.

The important bit is that file reading is performed on a separate
process to the renderer, and that that's set up in a secure way. I think
I got it right but that bit's not actually working right now so who
knows. The importanter bit is that I get something that's not just error
messages in the Electron pane: in this case, it's the skeleton of the
Editor component. So, effectively, strip away all the Electron stuff and
this commit just introduces a dirty hotfix to the identifier function.