Add file path to response channel name for files
I noticed that, when reading the contents of files transmitted over the
IPC from the file read/write channel in the Electron preload script,
even though I had two different events defined, they were both
identical. I was able to work out that the file contents were likely
being contaminated on account of sharing a data channel. By including
the file path in the data channel name as a unique indentifier, I can
guarantee that the read event for a file should go to the right place.

Curiously, I was already using one-time response events to communicate
with the Electron backend, so I would've expected this bug to result in
events being in the wrong order rather than being duplicated.