Implement EditConditional
This commit gets as far as creating the EditConditional component and
seeing if its core conceit - that is, attaching the wanted component as
a dynamic component - works. It does. But there's still a lot of work to
be done, including creating dynamic components for each term of the
conditional and propagating events back upwards which at the moment is
entirely unimplemented.