Set Identifier to string
It was previously string[], e.g. ["eventId", "interactionId"] would have
pointed to a certain interaction in a certain event. I did it this way
in order to have the possibility of type-checking a statically-typed
config file by the TypeScript interpreter.

However, I no longer have a statically-typed config file, so this is
pointless and brings more problems than it's worth. Complex identifier
that must point outside their common scope can simply have a dot in them
or something.