Apply conditions and conditionals to the schema
In the new flattened database system, of the things that conditions used
to be able to apply to, they can now only apply to three: options,
messageGroups, and messages. They will also be able to apply to actions,
but I've not gotten to these yet.

In the old system, almost anything could be a conditional. (Terminology
reminder: a 'conditional' is a static data structure containing if, elif
and else branches, and at runtime resolves to a single value) In the
new system, few things can be conditional, but the definition of
conditional is slightly expanded: a conditional can now resolve to any
number of values, including 0. Currently, only messages and
messageGroups can be conditional; this is where they will be most used
and I feel that implementing them in too many other places would
introduce a huge amount of UI complexity.

Options are not conditional per the definition above; however, they have
a single condition applied to them that determines whether or not they
appear. The condition cannot be null, but it may have zero terms (which
is equvalent to true).

messageGroups, messages, and eventually actions are conditional. The way
I have implemented this is implicit and therefore hard to reason about,
but it is fairly elegant, and it does mesh very well with the flat
database structure.

Each item of those tables has a condition column.  The conditional is
made up of a series of condition IDs, where NULL indicates an 'else'
value. The 'else' value must come first (ordering by sortIndex), even
though it will be displayed last in the UI. The next non-null condition
ID is the 'if' condition, and its row contains the item for the 'if'
branch. Any subsequent non-NULL condition IDs are 'elif' branches.

There can be multiple items in a single conditional branch; these are
indicated by sharing a condition ID. Everything else would have a unique
condition ID created for it.

A lone 'else' conditional branch not followed immediately (in terms of
sortIndex) by a row with a non-NULL condition ID indicates a
non-conditional value. For the data types that can be conditional, this
should be displayed as just a normal value with no condition associated
with it. As a result, all values of these three types are always
conditional, although most of them will have a condition with zero
terms.

Overall, the implementation of the conditional data structure in terms
of the database schema is a little convoluted, it is not explicit, and
it makes improper use of the NULL value. But it only takes up a single
column for each data type that needs it, and preserves strong database
integrity (as opposed to having a single table contain all conditionals,
which would've been a nightmare in terms of foreign key relations).

With proper documentation, there should be no issues.