Track Conditional root ID
Nodes that can be Conditional (i.e. in an if-elif-else structure)
(presently just messageGroups and messages) now track which node is the
'root' of that Conditional. The root is the first item in the 'else'
branch. This item is guaranteed to exist: lone, apparently
non-conditional items are represented by a single 'else'; ordinary
Conditionals always have an 'else'; and Conditionals that appear to only
have 'if' and 'elif' branches actually do have an 'else' that returns
nothing.

This is necessary because the previous solution (using NULL to
distinguish else branches) works well for if and elif but fails to
distinguish between a Conditional with an else branch containing two
items and a Conditional with an else branch containing one item followed
by a subsequent lone item.

Tracking which Conditional an item is part of solves this.

Lone items will track themselves as their own Conditional root.