Add check for root ID position
The first item in any Conditional is the first item of the 'else'
branch, even if that Conditional does not have any further items. As a
result, the first item in a set that can contain Conditionals must
always have a Conditional root ID pointing to itself.

If there are other Conditionals in the set, there will be other items
whose root ID is themself, however these cannot be trivially checked for
at the database level.

The not equals operator is being used as XOR.