Create TypeScript types for database entries
This commit replaces the old TS types with a new type that would
represent the database schema as a TS object. This would be used as the
basis for Vue (likely Vuex) reactivity.

The old types contained a few things that were specific to a single
Convomap consumer, which will need to be reimplemented in that project's
runtime: message timing control and delays.

Additionally, the old types contained a base set of imperative actions
and conditions. These have been removed and will eventually return as
default action functions and condition functions, added automatically to
new Convomap projects.