Convert JSON sample to SQL sample
SQLite databases and optionally SQL dumps are going to be used for
storing Convomap data, so I am ditching the JSON sample after converting
it to SQL.

The sample_events dir has also been renamed to samples.

Both the default DB storage method (.convomap) and the SQL dump method
(.convomap.sql) are stored as an example of what the output will look
like.